+++ CHANGELOG.md
@@ -86,6 +86,7 @@
 - PIM-10033: Implement LRU cache for Asset Families to optimize performance
 - PIM-10055: Optimize publication of incoming published asssocication upon product publication
 - PIM-10063: Fix product link rule execution to make the property case insensitive
+- PIM-10059: Fix completeness display when user does not have permission on attribute group(s)
 
 ## Improvements